начали выносить movie в отдельный презентор
